@@ -44,97 +40,141 @@ group = "com.mod-buildcraft-compat"
 archivesBaseName = "buildcraft" // the name that all artifacts will use as a base. artifacts names follow this pattern: [baseName]-[appendix]-[version]-[classifier].[extension]
 
 ext {
-    compatModInfo = new groovy.json.JsonSlurper().parse(file("resources/mcmod.info"))
-    mainModInfo = new groovy.json.JsonSlurper().parse(file("BuildCraft/buildcraft_resources/mcmod.info"))
+//    compatModInfo = file("resources/mods.toml")
+//    mainModInfo = file("BuildCraft/buildcraft_resources/mods.toml")
+    compatModInfo = file("mod_info/META-INF/mods.toml").readLines()
+    mainModInfo = file("BuildCraft/mod_info/META-INF/mods.toml").readLines()
 }
 
-sourceCompatibility = 1.8
-targetCompatibility = 1.8
+// Mojang ships Java 17 to end users in 1.18+, so your mod should target Java 17.
+java.toolchain.languageVersion = JavaLanguageVersion.of(17)
 
-repositories {
-    maven { name="JEI"; url="http://dvs1.progwml6.com/files/maven" }
-    maven { name="TOP"; url="http://maven.tterrag.com/" }
-    maven { name="hwyla"; url="http://tehnut.info/maven" }
-    // The One probe depends on the RF API -- we don't use it ourselves, nor will we ever add compat for it...
-    maven { name="cofh"; url="http://maven.covers1624.net" }
-    maven { name="forestry"; url = "http://maven.ic2.player.to/" }
-    maven { name "CraftTweaker"; url "http://maven.blamejared.com/" }
-}
-
-dependencies {
-    // JEI Stuff
-    // Versions from http://minecraft.curseforge.com/projects/just-enough-items-jei/files
-    compileOnly "mezz.jei:jei_${config.mc_version}:${config.jei_version}:api"
-    runtime "mezz.jei:jei_${config.mc_version}:${config.jei_version}"
-
-    // The One Probe Stuff
-    compileOnly "mcjty.theoneprobe:TheOneProbe-1.12:${config.top_version}:api"
-    runtime "mcjty.theoneprobe:TheOneProbe-1.12:${config.top_version}"
-    // The One Probe depends on the RF api for some odd reason... we never intend to add support for RF ourselves though.
-
-    // HWYLA
-    compileOnly "mcp.mobius.waila:Hwyla:${config.hwyla_version}:api"
-    runtime "mcp.mobius.waila:Hwyla:${config.hwyla_version}"
-
-    // Forestry
-    deobfCompile "net.sengir.forestry:forestry_${config.mc_version}:${config.forestry_version}"
-
-    deobfCompile "net.industrial-craft:industrialcraft-2:${config.ic2_version}"
-
-    // CraftTweaker
-    deobfCompile "CraftTweaker2:CraftTweaker2-API:${config.crafttweaker_version}"
-    deobfCompile "CraftTweaker2:CraftTweaker2-MC1120-Main:1.12-${config.crafttweaker_version}"
-}
-
+println "Java: ${System.getProperty 'java.version'}, JVM: ${System.getProperty 'java.vm.version'} (${System.getProperty 'java.vendor'}), Arch: ${System.getProperty 'os.arch'}"
 minecraft {
-    version = config.mc_version + "-" + config.forge_version
-    runDir = "run"
+    // The mappings can be changed at any time and must be in the following format.
+    // Channel:   Version:
+    // official   MCVersion             Official field/method names from Mojang mapping files
+    // parchment  YYYY.MM.DD-MCVersion  Open community-sourced parameter names and javadocs layered on top of official
+    //
+    // You must be aware of the Mojang license when using the 'official' or 'parchment' mappings.
+    // See more information here: https://github.com/MinecraftForge/MCPConfig/blob/master/Mojang.md
+    //
+    // Parchment is an unofficial project maintained by ParchmentMC, separate from MinecraftForge
+    // Additional setup is needed to use their mappings: https://parchmentmc.org/docs/getting-started
+    //
+    // Use non-default mappings at your own risk. They may not always work.
+    // Simply re-run your setup task after changing the mappings to update your workspace.
+    mappings channel: config.mappings_channel, version: config.mappings_version
 
-    // the mappings can be changed at any time, and must be in the following format.
-    // snapshot_YYYYMMDD   snapshot are built nightly.
-    // stable_#            stables are built at the discretion of the MCP team.
-    // Use non-default mappings at your own risk. they may not allways work.
-    // simply re-run your setup task after changing the mappings to update your workspace.
-    mappings = config.mappings_version
+    // When true, this property will have all Eclipse/IntelliJ IDEA run configurations run the "prepareX" task for the given run configuration before launching the game.
+    // In most cases, it is not necessary to enable.
+    // enableEclipsePrepareRuns = true
+    // enableIdeaPrepareRuns = true
 
-    useDepAts = true
+    // This property allows configuring Gradle's ProcessResources task(s) to run on IDE output locations before launching the game.
+    // It is REQUIRED to be set to true for this template to function.
+    // See https://docs.gradle.org/current/dsl/org.gradle.language.jvm.tasks.ProcessResources.html
+    copyIdeResources = true
+
+    // When true, this property will add the folder name of all declared run configurations to generated IDE run configurations.
+    // The folder name can be set on a run configuration using the "folderName" property.
+    // By default, the folder name of a run configuration is the name of the Gradle project containing it.
+    // generateRunFolders = true
+
+    // This property enables access transformers for use in development.
+    // They will be applied to the Minecraft artifact.
+    // The access transformer file can be anywhere in the project.
+    // However, it must be at "META-INF/accesstransformer.cfg" in the final mod jar to be loaded by Forge.
+    // This default location is a best practice to automatically put the file in the right place in the final jar.
+    // See https://docs.minecraftforge.net/en/latest/advanced/accesstransformers/ for more information.
+    accessTransformer = file('BuildCraft/buildcraft_resources/META-INF/accesstransformer.cfg')
 
     def separate = Boolean.getBoolean("build_compat_only")
     if (separate) {
         project.version = config.compat_version
     } else {
-        project.version = config.mod_version;
+//        project.version = config.mod_version;
+        project.version = config.compat_version + "-" + config.mc_version + "-" + new Date().format("yyyy-MM-dd'T'HH-mm-ssZ")
     }
 
-    def addVersionDetails = !Boolean.getBoolean("release")
+//    def addVersionDetails = !Boolean.getBoolean("release")
+//
+//    // Git versioning stuffs
+//    if (grgit != null) {
+//        def repo = grgit.open(dir: project.rootDir)
+//        replace "\${git_commit_hash}", repo.head().id
+//        replace "\${git_commit_msg}", repo.head().fullMessage.replace("\"", "\\\\\\\"").split("\n")[0]
+//        replace "\${git_commit_author}", repo.head().author.name
+//        replace "\${git_branch}", repo.branch.current().getName()
+//        if (addVersionDetails) {
+//            project.version += "-" + repo.head().id.toString().substring(0, 12)
+//        }
+//    } else {
+//        if (addVersionDetails) {
+//            project.version += "-SNAPSHOT"
+//        }
+//    }
+//
+//    // replacing stuff in the source
+//    replace "\$version", project.version
+//    replace "\${mcversion}", version
+//    // Replace @Mod.acceptedMinecraftVersions specially as it has to be a valid version in a dev environment :/
+//    replace "(gradle_replace_mcversion,)", "[" + version + "]"
+//    replace "(gradle_replace_forgeversion,)", "[" + config.forge_version + ",)"
+//    replace "\$bc_version", config.mod_version
 
-    // Git versioning stuffs
-    if (grgit != null) {
-        def repo = grgit.open(dir: project.rootDir)
-        replace "\${git_commit_hash}", repo.head().id
-        replace "\${git_commit_msg}", repo.head().fullMessage.replace("\"", "\\\\\\\"").split("\n")[0]
-        replace "\${git_commit_author}", repo.head().author.name
-        replace "\${git_branch}", repo.branch.current().getName()
-        if (addVersionDetails) {
-            project.version += "-" + repo.head().id.toString().substring(0, 12)
+    // Default run configurations.
+    // These can be tweaked, removed, or duplicated as needed.
+    runs {
+        // applies to all the run configs below
+        configureEach {
+            workingDirectory project.file('run')
+
+            // Recommended logging data for a userdev environment
+            // The markers can be added/remove as needed separated by commas.
+            // "SCAN": For mods scan.
+            // "REGISTRIES": For firing of registry events.
+            // "REGISTRYDUMP": For getting the contents of all registries.
+            property 'forge.logging.markers', 'REGISTRIES'
+
+            // Recommended logging level for the console
+            // You can set various levels here.
+            // Please read: https://stackoverflow.com/questions/2031163/when-to-use-the-different-log-levels
+            property 'forge.logging.console.level', 'debug'
+
+            mods {
+                "${mod_id}" {
+                    source sourceSets.main
+                }
+            }
         }
-    } else {
-        if (addVersionDetails) {
-            project.version += "-SNAPSHOT"
+
+        client {
+            // Comma-separated list of namespaces to load gametests from. Empty = all namespaces.
+            property 'forge.enabledGameTestNamespaces', mod_id
+        }
+
+        server {
+            workingDirectory project.file('run_server')
+            property 'forge.enabledGameTestNamespaces', mod_id
+            args '--nogui'
+        }
+
+        // This run config launches GameTestServer and runs all registered gametests, then exits.
+        // By default, the server will crash when no gametests are provided.
+        // The gametest system is also enabled by default for other run configs under the /test command.
+        gameTestServer {
+            property 'forge.enabledGameTestNamespaces', mod_id
+        }
+
+        data {
+            // example of overriding the workingDirectory set in configureEach above
+            workingDirectory project.file('run-data')
+
+            // Specify the modid for data generation, where to output the resulting resource, and where to look for existing resources.
+            args '--mod', mod_id, '--all', '--output', file('BuildCraft/buildcraft_resources_generated/'), '--existing', file('BuildCraft/buildcraft_resources/')
         }
     }
-
-    // makeObfSourceJar = false // an Srg named sources jar is made by default. uncomment this to disable.
-
-    // replacing stuff in the source
-    replace "\$version", project.version
-    replace "\${mcversion}", version
-    // Replace @Mod.acceptedMinecraftVersions specially as it has to be a valid version in a dev environment :/
-    replace "(gradle_replace_mcversion,)", "[" + version + "]"
-    replace "(gradle_replace_forgeversion,)", "[" + config.forge_version + ",)"
-    replace "\$bc_version", config.mod_version
-
-    useDepAts = true
 }
 
 compileJava {
