@@ -1,10 +1,11 @@
 package buildcraft.compat;
 
-import net.minecraftforge.fml.common.Loader;
+import net.minecraftforge.fml.ModList;
 
 public abstract class CompatModuleBase {
     public boolean canLoad() {
-        return Loader.isModLoaded(this.compatModId());
+//        return Loader.isModLoaded(this.compatModId());
+        return ModList.get().isLoaded(this.compatModId());
     }
 
     public abstract String compatModId();
