@@ -1,29 +1,29 @@
 package buildcraft.compat;
 
+import buildcraft.compat.network.IGuiCreator;
+import buildcraft.lib.misc.StackUtil;
+import buildcraft.lib.tile.item.ItemHandlerSimple;
+import com.google.common.collect.Lists;
+import net.minecraft.world.item.ItemStack;
+import net.minecraftforge.common.capabilities.Capability;
+import net.minecraftforge.common.capabilities.CapabilityManager;
+import net.minecraftforge.common.capabilities.CapabilityToken;
+import net.minecraftforge.common.capabilities.RegisterCapabilitiesEvent;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.minecraftforge.fml.common.Mod;
+
 import java.util.List;
 
-import com.google.common.collect.Lists;
-
-import net.minecraft.item.ItemStack;
-
-import net.minecraftforge.common.capabilities.Capability;
-
-import buildcraft.api.core.CapabilitiesHelper;
-
-import buildcraft.lib.tile.item.ItemHandlerSimple;
-
-import buildcraft.compat.network.IGuiCreator;
-
+@Mod.EventBusSubscriber(bus = Mod.EventBusSubscriber.Bus.MOD, modid = BCCompat.MODID)
 public class CompatUtils {
+    public static final Capability<IGuiCreator> CAP_GUI_CREATOR = CapabilityManager.get(new CapabilityToken<>() {
+    });
 
-    public static final Capability<IGuiCreator> CAP_GUI_CREATOR =
-        CapabilitiesHelper.registerCapability(IGuiCreator.class);
-
-    private CompatUtils() {}
+    private CompatUtils() {
+    }
 
     public static List<ItemStack> compactInventory(ItemHandlerSimple inventory) {
         List<ItemStack> stacks = Lists.newArrayList();
-
         for (int slot = 0; slot < inventory.getSlots(); slot++) {
             ItemStack stack = inventory.getStackInSlot(slot);
             if (stack.isEmpty()) {
@@ -32,7 +32,8 @@ public class CompatUtils {
 
             boolean handled = false;
             for (ItemStack existing : stacks) {
-                if (existing.isItemEqual(stack)) {
+//                if (existing.isItemEqual(stack))
+                if (StackUtil.isSameItemSameDamage(existing, stack)) {
                     existing.grow(stack.getCount());
                     handled = true;
                     break;
@@ -44,5 +45,11 @@ public class CompatUtils {
         }
 
         return stacks;
+    }
+
+    // Calen
+    @SubscribeEvent
+    public static void registerCapability(RegisterCapabilitiesEvent event) {
+        event.register(IGuiCreator.class);
     }
 }
