@@ -2,5 +2,6 @@
 @MethodsReturnNonnullByDefault
 package buildcraft.compat.module.jei.energy.combustionengine;
 
+import net.minecraft.MethodsReturnNonnullByDefault;
+
 import javax.annotation.ParametersAreNonnullByDefault;
-import mcp.MethodsReturnNonnullByDefault;
