@@ -1,34 +1,51 @@
 package buildcraft.compat.module.jei.transferhandlers;
 
-import javax.annotation.Nullable;
-import net.minecraft.entity.player.EntityPlayer;
+import buildcraft.silicon.BCSiliconMenuTypes;
 import buildcraft.silicon.container.ContainerAdvancedCraftingTable;
-import mezz.jei.api.gui.IRecipeLayout;
+import mezz.jei.api.constants.RecipeTypes;
+import mezz.jei.api.gui.ingredient.IRecipeSlotsView;
+import mezz.jei.api.recipe.RecipeType;
 import mezz.jei.api.recipe.transfer.IRecipeTransferError;
 import mezz.jei.api.recipe.transfer.IRecipeTransferHandler;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.inventory.MenuType;
+import net.minecraft.world.item.crafting.CraftingRecipe;
 
-public class AdvancedCraftingItemsTransferHandler implements IRecipeTransferHandler<ContainerAdvancedCraftingTable> {
+import javax.annotation.Nullable;
+import java.util.Optional;
+
+public class AdvancedCraftingItemsTransferHandler implements IRecipeTransferHandler<ContainerAdvancedCraftingTable, CraftingRecipe> {
+    public AdvancedCraftingItemsTransferHandler() {
+    }
+
     @Override
     public Class<ContainerAdvancedCraftingTable> getContainerClass() {
         return ContainerAdvancedCraftingTable.class;
     }
 
-    @Nullable
     @Override
-    public IRecipeTransferError transferRecipe(ContainerAdvancedCraftingTable container, IRecipeLayout recipeLayout, EntityPlayer player, boolean maxTransfer, boolean doTransfer) {
-        if (doTransfer) {
-//            Map<Integer, ? extends IGuiIngredient<ItemStack>> inputs = recipeLayout.getItemStacks().getGuiIngredients();
-//
-//            for (int slot = 0; slot < 9; slot++) {
-//                IGuiIngredient<ItemStack> ingredient = inputs.getOrDefault(slot + 1, null);
-//                ItemStack stack = (ingredient == null) ? ItemStack.EMPTY : ingredient.getDisplayedIngredient();
-//
-//                container.sendSetPhantomSlot(container.tile.invBlueprint, slot, (stack == null) ? ItemStack.EMPTY : stack);
-//            }
+    public Optional<MenuType<ContainerAdvancedCraftingTable>> getMenuType() {
+        return Optional.of(BCSiliconMenuTypes.ADVANCED_CRAFTING_TABLE);
+    }
 
-            AutoCraftItemsTransferHandler.transferRecipe(
-                    itemStacks -> container.sendSetPhantomSlots(container.tile.invBlueprint, itemStacks),
-                    recipeLayout);
+    @Override
+    public RecipeType<CraftingRecipe> getRecipeType() {
+        return RecipeTypes.CRAFTING;
+    }
+
+    @Override
+    @Nullable
+//    public IRecipeTransferError transferRecipe(ContainerAdvancedCraftingTable container, IRecipeLayout recipeLayout, Player player, boolean maxTransfer, boolean doTransfer)
+    public IRecipeTransferError transferRecipe(ContainerAdvancedCraftingTable container, CraftingRecipe recipe, IRecipeSlotsView recipeSlots, Player player, boolean maxTransfer, boolean doTransfer) {
+        if (doTransfer) {
+            AutoCraftItemsTransferHandler.transferRecipe((itemStacks) ->
+                    {
+                        container.sendSetPhantomSlots(container.tile.invBlueprint, itemStacks);
+                    },
+//                    recipeLayout
+                    recipe,
+                    recipeSlots
+            );
         }
 
         return null;
