@@ -1,53 +1,74 @@
 package buildcraft.compat.module.jei.transferhandlers;
 
-import java.util.List;
-import java.util.Map;
-import java.util.function.Consumer;
-import javax.annotation.Nullable;
-import com.google.common.collect.Lists;
-import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.item.ItemStack;
+import buildcraft.factory.BCFactoryMenuTypes;
 import buildcraft.factory.container.ContainerAutoCraftItems;
-import mezz.jei.api.gui.IGuiIngredient;
-import mezz.jei.api.gui.IRecipeLayout;
+import com.google.common.collect.Lists;
+import mezz.jei.api.constants.RecipeTypes;
+import mezz.jei.api.gui.ingredient.IRecipeSlotView;
+import mezz.jei.api.gui.ingredient.IRecipeSlotsView;
+import mezz.jei.api.recipe.RecipeType;
 import mezz.jei.api.recipe.transfer.IRecipeTransferError;
 import mezz.jei.api.recipe.transfer.IRecipeTransferHandler;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.inventory.MenuType;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.crafting.CraftingRecipe;
 
-public class AutoCraftItemsTransferHandler implements IRecipeTransferHandler<ContainerAutoCraftItems> {
+import javax.annotation.Nullable;
+import java.util.List;
+import java.util.Optional;
+import java.util.function.Consumer;
+
+public class AutoCraftItemsTransferHandler implements IRecipeTransferHandler<ContainerAutoCraftItems, CraftingRecipe> {
+    public AutoCraftItemsTransferHandler() {
+    }
+
     @Override
     public Class<ContainerAutoCraftItems> getContainerClass() {
         return ContainerAutoCraftItems.class;
     }
 
-    @Nullable
     @Override
-    public IRecipeTransferError transferRecipe(ContainerAutoCraftItems container, IRecipeLayout recipeLayout, EntityPlayer player, boolean maxTransfer, boolean doTransfer) {
-        if (doTransfer) {
-//            Map<Integer, ? extends IGuiIngredient<ItemStack>> inputs = recipeLayout.getItemStacks().getGuiIngredients();
-//
-//            for (int slot = 0; slot < 9; slot++) {
-//                IGuiIngredient<ItemStack> ingredient = inputs.getOrDefault(slot + 1, null);
-//                ItemStack stack = (ingredient == null) ? ItemStack.EMPTY : ingredient.getDisplayedIngredient();
-//
-//                container.sendSetPhantomSlot(container.tile.invBlueprint, slot, (stack == null) ? ItemStack.EMPTY : stack);
-//            }
+    public Optional<MenuType<ContainerAutoCraftItems>> getMenuType() {
+        return Optional.of(BCFactoryMenuTypes.AUTO_WORKBENCH_ITEMS);
+    }
 
-            AutoCraftItemsTransferHandler.transferRecipe(
-                    itemStacks -> container.sendSetPhantomSlots(container.tile.invBlueprint, itemStacks),
-                    recipeLayout);
+    @Override
+    public RecipeType<CraftingRecipe> getRecipeType() {
+        return RecipeTypes.CRAFTING;
+    }
+
+    @Override
+    @Nullable
+//    public IRecipeTransferError transferRecipe(ContainerAutoCraftItems container, IRecipeLayout recipeLayout, Player player, boolean maxTransfer, boolean doTransfer)
+    public IRecipeTransferError transferRecipe(ContainerAutoCraftItems container, CraftingRecipe recipe, IRecipeSlotsView recipeSlots, Player player, boolean maxTransfer, boolean doTransfer) {
+        if (doTransfer) {
+            transferRecipe((itemStacks) ->
+                    {
+                        container.sendSetPhantomSlots((container.tile).invBlueprint, itemStacks);
+                    },
+//                    recipeLayout
+                    recipe,
+                    recipeSlots
+            );
         }
 
         return null;
     }
 
-    static void transferRecipe(Consumer<List<ItemStack>> callback, IRecipeLayout recipeLayout) {
-        Map<Integer, ? extends IGuiIngredient<ItemStack>> inputs = recipeLayout.getItemStacks().getGuiIngredients();
+    // static void transferRecipe(Consumer<List<ItemStack>> callback, IRecipeLayout recipeLayout)
+    static void transferRecipe(Consumer<List<ItemStack>> callback, CraftingRecipe recipe, IRecipeSlotsView recipeLayout) {
+//        Map<Integer, ? extends IGuiIngredient<ItemStack>> inputs = recipeLayout.getItemStacks().getGuiIngredients();
 
+        List<IRecipeSlotView> inputs = recipeLayout.getSlotViews();
         List<ItemStack> stacks = Lists.newArrayList();
-        for (int slot = 0; slot < 9; slot++) {
-            IGuiIngredient<ItemStack> ingredient = inputs.getOrDefault(slot + 1, null);
-            ItemStack stack = (ingredient == null) ? ItemStack.EMPTY : ingredient.getDisplayedIngredient();
-            stacks.add((stack == null) ? ItemStack.EMPTY : stack);
+
+        for (int slot = 0; slot < 9; ++slot) {
+//            IGuiIngredient<ItemStack> ingredient = (IGuiIngredient) inputs.getOrDefault(slot + 1, (Object) null);
+            IRecipeSlotView ingredient = inputs.get(slot + 1);
+//            ItemStack stack = ingredient == null ? ItemStack.EMPTY : (ItemStack) ingredient.getDisplayedIngredient();
+            ItemStack stack = ingredient == null ? ItemStack.EMPTY : (ItemStack) ingredient.getDisplayedIngredient().get().getIngredient();
+            stacks.add(stack == null ? ItemStack.EMPTY : stack);
         }
 
         callback.accept(stacks);
