@@ -1,72 +1,62 @@
 package buildcraft.compat.module.theoneprobe;
 
-import static buildcraft.compat.module.theoneprobe.BCPluginTOP.TOP_MOD_ID;
-
-import java.util.List;
-
-import com.google.common.base.Function;
-
-import net.minecraft.block.state.IBlockState;
-import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.item.ItemStack;
-import net.minecraft.tileentity.TileEntity;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.text.TextFormatting;
-import net.minecraft.world.World;
-
-import net.minecraftforge.fml.common.Loader;
-import net.minecraftforge.fml.common.Optional;
-import net.minecraftforge.fml.common.event.FMLInterModComms;
-
 import buildcraft.api.BCModules;
 import buildcraft.api.mj.ILaserTarget;
 import buildcraft.api.mj.MjAPI;
-
-import buildcraft.lib.tile.craft.IAutoCraft;
-
 import buildcraft.compat.CompatUtils;
+import buildcraft.lib.misc.BlockUtil;
+import buildcraft.lib.tile.craft.IAssemblyCraft;
+import buildcraft.lib.tile.craft.IAutoCraft;
+import com.google.common.base.Function;
+import mcjty.theoneprobe.api.*;
+import net.minecraft.network.chat.Component;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraft.world.level.block.state.BlockState;
 
-import mcjty.theoneprobe.api.ElementAlignment;
-import mcjty.theoneprobe.api.IBlockDisplayOverride;
-import mcjty.theoneprobe.api.IProbeHitData;
-import mcjty.theoneprobe.api.IProbeInfo;
-import mcjty.theoneprobe.api.IProbeInfoProvider;
-import mcjty.theoneprobe.api.ITheOneProbe;
-import mcjty.theoneprobe.api.ProbeMode;
+import java.util.List;
 
-@Optional.InterfaceList({
-        @Optional.Interface(modid = TOP_MOD_ID, iface = "mcjty.theoneprobe.api.IBlockDisplayOverride"),
-        @Optional.Interface(modid = TOP_MOD_ID, iface = "mcjty.theoneprobe.api.IProbeInfoProvider")
-})
-public class BCPluginTOP implements Function<ITheOneProbe, Void>, IBlockDisplayOverride, IProbeInfoProvider {
+//@InterfaceList({@Interface(
+//        modid = "theoneprobe",
+//        iface = "mcjty.theoneprobe.api.IBlockDisplayOverride"
+//), @Interface(
+//        modid = "theoneprobe",
+//        iface = "mcjty.theoneprobe.api.IProbeInfoProvider"
+//)})
+//public class BCPluginTOP implements Function<ITheOneProbe, Void>, IBlockDisplayOverride, IProbeInfoProvider
+public enum BCPluginTOP implements Function<ITheOneProbe, Void>, IBlockDisplayOverride, IProbeInfoProvider {
+    INSTANCE;
     static final String TOP_MOD_ID = "theoneprobe";
 
-    @Override
-    @Optional.Method(modid = TOP_MOD_ID)
+    // @Method(modid = "theoneprobe")
     public Void apply(ITheOneProbe top) {
         top.registerBlockDisplayOverride(this);
         top.registerProvider(this);
         return null;
     }
 
-    @Override
-    @Optional.Method(modid = TOP_MOD_ID)
-    public boolean overrideStandardInfo(ProbeMode mode, IProbeInfo probeInfo, EntityPlayer player, World world, IBlockState blockState, IProbeHitData data) {
+    // @Method(modid = "theoneprobe")
+    // public boolean overrideStandardInfo(ProbeMode mode, IProbeInfo probeInfo, EntityPlayer player, World world, IBlockState blockState, IProbeHitData data)
+    public boolean overrideStandardInfo(ProbeMode mode, IProbeInfo probeInfo, Player player, Level world, BlockState blockState, IProbeHitData data) {
         return false;
     }
 
-    @Override
-    @Optional.Method(modid = TOP_MOD_ID)
-    public String getID() {
-        return "buildcraftcompat.top";
+    // @Method(modid = "theoneprobe")
+    // public String getID()
+    public ResourceLocation getID() {
+//        return "buildcraftcompat.top";
+        return new ResourceLocation("buildcraftcompat.top");
     }
 
-    @Override
-    @Optional.Method(modid = TOP_MOD_ID)
-    public void addProbeInfo(ProbeMode mode, IProbeInfo probeInfo, EntityPlayer player, World world, IBlockState blockState, IProbeHitData data) {
-        ResourceLocation blockRegistryName = blockState.getBlock().getRegistryName();
-        if ((blockRegistryName != null) && (BCModules.isBcMod(blockRegistryName.getResourceDomain()))) {
-            TileEntity entity = world.getTileEntity(data.getPos());
+    // @Method(modid = "theoneprobe")
+//    public void addProbeInfo(ProbeMode mode, IProbeInfo probeInfo, EntityPlayer player, World world, IBlockState blockState, IProbeHitData data)
+    public void addProbeInfo(ProbeMode mode, IProbeInfo probeInfo, Player player, Level world, BlockState blockState, IProbeHitData data) {
+        ResourceLocation blockRegistryName = BlockUtil.getRegistryName(blockState.getBlock());
+        if (blockRegistryName != null && BCModules.isBcMod(blockRegistryName.getNamespace())) {
+            BlockEntity entity = world.getBlockEntity(data.getPos());
             if (entity instanceof IAutoCraft) {
                 this.addAutoCraftInfo(probeInfo, (IAutoCraft) entity);
             }
@@ -74,36 +64,50 @@ public class BCPluginTOP implements Function<ITheOneProbe, Void>, IBlockDisplayO
             if (entity instanceof ILaserTarget) {
                 this.addLaserTargetInfo(probeInfo, (ILaserTarget) entity);
             }
+
+            if (entity instanceof IAssemblyCraft) {
+                this.addAssemblyInfo(probeInfo, (IAssemblyCraft) entity);
+            }
         }
+
     }
 
-    @Optional.Method(modid = TOP_MOD_ID)
+    // @Method(modid = "theoneprobe")
     private void addAutoCraftInfo(IProbeInfo probeInfo, IAutoCraft crafter) {
         if (!crafter.getCurrentRecipeOutput().isEmpty()) {
             IProbeInfo mainInfo = probeInfo.vertical();
-            mainInfo
-                    .horizontal(mainInfo.defaultLayoutStyle().alignment(ElementAlignment.ALIGN_CENTER))
-                    .text("Making: ")
-                    .item(crafter.getCurrentRecipeOutput());
-            IProbeInfo info = mainInfo
-                    .horizontal(mainInfo.defaultLayoutStyle().alignment(ElementAlignment.ALIGN_CENTER))
-                    .text("From: ");
+//            mainInfo.horizontal(mainInfo.defaultLayoutStyle().alignment(ElementAlignment.ALIGN_CENTER)).text("Making: ").item(crafter.getCurrentRecipeOutput());
+            mainInfo.horizontal(mainInfo.defaultLayoutStyle().alignment(ElementAlignment.ALIGN_CENTER)).text(Component.translatable("buildcraft.waila.crafting")).item(crafter.getCurrentRecipeOutput());
+//            IProbeInfo info = mainInfo.horizontal(mainInfo.defaultLayoutStyle().alignment(ElementAlignment.ALIGN_CENTER)).text("From: ");
+            IProbeInfo info = mainInfo.horizontal(mainInfo.defaultLayoutStyle().alignment(ElementAlignment.ALIGN_CENTER)).text(Component.translatable("buildcraft.waila.crafting_from"));
             List<ItemStack> stacks = CompatUtils.compactInventory(crafter.getInvBlueprint());
 
-            for (ItemStack stack : stacks)
+            for (ItemStack stack : stacks) {
                 info.item(stack);
+            }
+        } else {
+            IProbeInfo mainInfo = probeInfo.vertical();
+            mainInfo.horizontal(mainInfo.defaultLayoutStyle().alignment(ElementAlignment.ALIGN_CENTER)).text(Component.translatable("buildcraft.waila.no_recipe"));
+        }
+
+    }
+
+    // @Method(modid = "theoneprobe")
+    private void addLaserTargetInfo(IProbeInfo probeInfo, ILaserTarget laserTarget) {
+        long power = laserTarget.getRequiredLaserPower();
+        if (power > 0L) {
+//            probeInfo.horizontal().text(TextFormatting.WHITE + "Waiting from laser: ").text(TextFormatting.AQUA + MjAPI.formatMj(power)).text(TextFormatting.AQUA + "MJ");
+            probeInfo.horizontal().text(Component.translatable("buildcraft.waila.waiting_for_laser", MjAPI.formatMj(power)));
         }
     }
 
-    @Optional.Method(modid = TOP_MOD_ID)
-    private void addLaserTargetInfo(IProbeInfo probeInfo, ILaserTarget laserTarget) {
-        long power = laserTarget.getRequiredLaserPower();
-        if (power > 0) {
-            probeInfo.horizontal()
-                    .text(TextFormatting.WHITE + "Waiting from laser: ")
-                    .text(TextFormatting.AQUA + MjAPI.formatMj(power))
-                    .text(TextFormatting.AQUA + "MJ");
+    private void addAssemblyInfo(IProbeInfo probeInfo, IAssemblyCraft assembly) {
+        ItemStack result = assembly.getAssemblyResult();
+        if (!result.isEmpty()) {
+            probeInfo.horizontal().text(Component.translatable("buildcraft.waila.crafting")).item(result);
+        } else {
+            IProbeInfo mainInfo = probeInfo.vertical();
+            mainInfo.horizontal(mainInfo.defaultLayoutStyle().alignment(ElementAlignment.ALIGN_CENTER)).text(Component.translatable("buildcraft.waila.no_recipe"));
         }
     }
 }
-
