@@ -1,57 +1,36 @@
 package buildcraft.compat.module.waila;
 
-import static buildcraft.compat.module.waila.HWYLAPlugin.WAILA_MOD_ID;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.entity.BlockEntity;
+import snownee.jade.api.BlockAccessor;
+import snownee.jade.api.IBlockComponentProvider;
+import snownee.jade.api.IServerDataProvider;
+import snownee.jade.api.ITooltip;
+import snownee.jade.api.config.IPluginConfig;
 
-import java.util.List;
-import javax.annotation.Nonnull;
-import net.minecraft.entity.player.EntityPlayerMP;
-import net.minecraft.item.ItemStack;
-import net.minecraft.nbt.NBTTagCompound;
-import net.minecraft.tileentity.TileEntity;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.World;
-import net.minecraftforge.fml.common.Optional;
+//class BaseWailaDataProvider implements IWailaDataProvider
+public interface BaseWailaDataProvider {
 
-import mcp.mobius.waila.api.IWailaConfigHandler;
-import mcp.mobius.waila.api.IWailaDataAccessor;
-import mcp.mobius.waila.api.IWailaDataProvider;
+    static abstract class BodyProvider implements IBlockComponentProvider {
+        @Override
+        public void appendTooltip(ITooltip iTooltip, BlockAccessor blockAccessor, IPluginConfig iPluginConfig) {
+            getWailaBody(iTooltip, blockAccessor, iPluginConfig);
+        }
 
-@Optional.InterfaceList({
-        @Optional.Interface(modid = WAILA_MOD_ID, iface = "mcp.mobius.waila.api.IWailaDataProvider")
-})
-class BaseWailaDataProvider implements IWailaDataProvider {
-    @Nonnull
-    @Override
-    @Optional.Method(modid = WAILA_MOD_ID)
-    public ItemStack getWailaStack(IWailaDataAccessor accessor, IWailaConfigHandler config) {
-        return ItemStack.EMPTY;
+        abstract void getWailaBody(ITooltip iTooltip, BlockAccessor accessor, IPluginConfig iPluginConfig);
     }
 
-    @Nonnull
-    @Override
-    @Optional.Method(modid = WAILA_MOD_ID)
-    public List<String> getWailaHead(ItemStack itemStack, List<String> currentTip, IWailaDataAccessor accessor, IWailaConfigHandler config) {
-        return currentTip;
-    }
+//    static abstract class NBTProvider implements IServerDataProvider<BlockEntity>
+    static abstract class NBTProvider implements IServerDataProvider<BlockAccessor> {
+        @Override
+//        public void appendServerData(CompoundTag tag, ServerPlayer player, Level blockAccessor, BlockEntity blockEntity, boolean showDetails)
+        public void appendServerData(CompoundTag tag, BlockAccessor accessor) {
+            getNBTData(tag, accessor);
+        }
 
-    @Nonnull
-    @Override
-    @Optional.Method(modid = WAILA_MOD_ID)
-    public List<String> getWailaBody(ItemStack itemStack, List<String> currentTip, IWailaDataAccessor accessor, IWailaConfigHandler config) {
-        return currentTip;
-    }
-
-    @Nonnull
-    @Override
-    @Optional.Method(modid = WAILA_MOD_ID)
-    public List<String> getWailaTail(ItemStack itemStack, List<String> currentTip, IWailaDataAccessor accessor, IWailaConfigHandler config) {
-        return currentTip;
-    }
-
-    @Nonnull
-    @Override
-    @Optional.Method(modid = WAILA_MOD_ID)
-    public NBTTagCompound getNBTData(EntityPlayerMP player, TileEntity te, NBTTagCompound tag, World world, BlockPos pos) {
-        return tag;
+//        abstract void getNBTData(CompoundTag tag, ServerPlayer player, Level blockAccessor, BlockEntity blockEntity, boolean showDetails);
+        abstract void getNBTData(CompoundTag tag, BlockAccessor accessor);
     }
 }
