@@ -1,49 +1,42 @@
 package buildcraft.compat.module.waila;
 
-import static buildcraft.compat.module.waila.HWYLAPlugin.WAILA_MOD_ID;
-
-import net.minecraft.item.ItemStack;
-
-import net.minecraftforge.fml.common.Optional;
-
-import buildcraft.api.mj.ILaserTarget;
-
-import buildcraft.lib.tile.craft.IAutoCraft;
-
-import mcp.mobius.waila.api.IWailaDataProvider;
-import mcp.mobius.waila.api.IWailaPlugin;
-import mcp.mobius.waila.api.IWailaRegistrar;
-import mcp.mobius.waila.api.SpecialChars;
-import mcp.mobius.waila.api.WailaPlugin;
+import buildcraft.lib.block.BlockBCTile_Neptune;
+import buildcraft.lib.tile.TileBC_Neptune;
+import snownee.jade.api.*;
 
 @WailaPlugin
-@Optional.InterfaceList({
-        @Optional.Interface(modid = WAILA_MOD_ID, iface = "mcp.mobius.waila.api.IWailaPlugin")
-})
 public class HWYLAPlugin implements IWailaPlugin {
-    static final String WAILA_MOD_ID = "waila";
+    // Calen: in 1.18.2 we should use Block or TE class instead of interface here, it's too difficult to get all matched if allowing some BC modules absent,
+    // so just use TileBC_Neptune and BlockBCTile_Neptune
+    @Override
+    public void register(IWailaCommonRegistration registrar) {
+        IServerDataProvider<BlockAccessor> autoCraftNbtProvider = new AutoCraftDataProvider.NBTProvider();
+        IServerDataProvider<BlockAccessor> laserTargetNbtProvider = new LaserTargetDataProvider.NBTProvider();
+        IServerDataProvider<BlockAccessor> assemblyCraftNbtProvider = new AssemblyCraftDataProvider.NBTProvider();
+        IServerDataProvider<BlockAccessor> mjStorageNbtProvider = new MjStorageDataProvider.NBTProvider();
+
+        registrar.registerBlockDataProvider(autoCraftNbtProvider, TileBC_Neptune.class);
+        registrar.registerBlockDataProvider(laserTargetNbtProvider, TileBC_Neptune.class);
+        registrar.registerBlockDataProvider(assemblyCraftNbtProvider, TileBC_Neptune.class);
+        registrar.registerBlockDataProvider(mjStorageNbtProvider, TileBC_Neptune.class);
+    }
 
     @Override
-    public void register(IWailaRegistrar registrar) {
-        IWailaDataProvider autoCraftProvider = new AutoCraftDataProvider();
-        registrar.registerNBTProvider(autoCraftProvider, IAutoCraft.class);
-        registrar.registerBodyProvider(autoCraftProvider, IAutoCraft.class);
+    public void registerClient(IWailaClientRegistration registrar) {
+//        IComponentProvider autoCraftBodyProvider = new AutoCraftDataProvider.BodyProvider();
+//        IComponentProvider laserTargetBodyProvider = new LaserTargetDataProvider.BodyProvider();
+//        IComponentProvider assemblyCraftBodyProvider = new AssemblyCraftDataProvider.BodyProvider();
+        IBlockComponentProvider autoCraftBodyProvider = new AutoCraftDataProvider.BodyProvider();
+        IBlockComponentProvider laserTargetBodyProvider = new LaserTargetDataProvider.BodyProvider();
+        IBlockComponentProvider assemblyCraftBodyProvider = new AssemblyCraftDataProvider.BodyProvider();
+        IBlockComponentProvider mjStorageBodyProvider = new MjStorageDataProvider.BodyProvider();
 
-        IWailaDataProvider laserTargetProvider = new LaserTargetDataProvider();
-        registrar.registerNBTProvider(laserTargetProvider, ILaserTarget.class);
-        registrar.registerBodyProvider(laserTargetProvider, ILaserTarget.class);
-    }
-
-    static String getItemStackString(ItemStack stack) {
-        return getItemStackString(stack, "1");
-    }
-
-    private static String getItemStackString(ItemStack stack, String thing) {
-        // TODO: find out what that 'thing' really is
-        return SpecialChars.getRenderString("waila.stack", thing,
-                stack.getItem().getRegistryName().toString(),
-                String.valueOf(stack.getCount()),
-                String.valueOf(stack.getItemDamage())
-        );
+//        registrar.registerComponentProvider(autoCraftBodyProvider, TooltipPosition.BODY, BlockBCTile_Neptune.class);
+//        registrar.registerComponentProvider(laserTargetBodyProvider, TooltipPosition.BODY, BlockBCTile_Neptune.class);
+//        registrar.registerComponentProvider(assemblyCraftBodyProvider, TooltipPosition.BODY, BlockBCTile_Neptune.class);
+        registrar.registerBlockComponent(autoCraftBodyProvider, BlockBCTile_Neptune.class);
+        registrar.registerBlockComponent(laserTargetBodyProvider, BlockBCTile_Neptune.class);
+        registrar.registerBlockComponent(assemblyCraftBodyProvider, BlockBCTile_Neptune.class);
+        registrar.registerBlockComponent(mjStorageBodyProvider, BlockBCTile_Neptune.class);
     }
 }
